requirements.txt
tweepy>=4.14.0
google-genai>=1.0.0
